move PendingTrace into its own component
