+++ BUILD.bazel
@@ -15,6 +15,8 @@ cc_library(
         "src/opentracing_external.cpp",
         "src/parse_util.cpp",
         "src/parse_util.h",
+        "src/pending_trace.cpp",
+        "src/pending_trace.h",
         "src/sample.cpp",
         "src/sample.h",
         "src/sampling_mechanism.cpp",
